requirements.txt
pyinstaller==5.13.0
pysimplegui==4.60.5
pywin32==306
pystray==0.19.4
pillow==10.0.0